requirements.txt
apipkg==1.4
appnope==0.1.0
bottle==0.12.13
certifi==2018.1.18
chardet==3.0.4
decorator==4.1.2
execnet==1.5.0
flake8==3.5.0
gevent==1.2.2
gevent-websocket==0.10.1
greenlet==0.4.12
idna==2.6
ipython==6.2.1
ipython-genutils==0.2.0
jedi==0.11.0
mccabe==0.6.1
parso==0.1.0
pep8==1.7.1
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
py==1.5.2
pycodestyle==2.3.1
pyflakes==1.6.0
Pygments==2.2.0
pytest==3.2.5
pytest-cache==1.0
pytest-flake8==0.9.1
pytest-pep8==1.0.6
requests==2.18.4
simplegeneric==0.8.1
six==1.11.0
tornado==4.5.3
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
websocket-client==0.46.0
